Fixed codacy issues
